Refactor and change Shoot to CommandGroup
