+++ src/org/robockets/steamworks/commands/Shoot.java
@@ -1,37 +1,19 @@
 package org.robockets.steamworks.commands;
 
-import edu.wpi.first.wpilibj.command.Command;
-import org.robockets.commons.YAxisRelativeDirection;
-import org.robockets.steamworks.Robot;
+import edu.wpi.first.wpilibj.command.CommandGroup;
 
 /**
  * @author Jake Backer
  */
-public class Shoot extends Command {
+public class Shoot extends CommandGroup {
 
     public Shoot() {
-
-    }
-
-    protected void initialize() {
-
-    }
-
-    protected void execute() {
-        Robot.shooter.spinUp();
-        Robot.conveyor.moveConveyor(YAxisRelativeDirection.FORWARD);
-    }
-
-    protected boolean isFinished() {
-        return false;
-    }
-
-    protected void end() {
-        Robot.shooter.stop();
-        Robot.conveyor.stop();
-    }
-
-    protected void interrupted() {
-        end();
+        addParallel(new SpinSpinners());
+        try {
+            wait(100); // This will be changed
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+        addSequential(new MoveConveyor(2)); // This will also be changed
     }
 }
